feat: setup project scaffolding
The structure of project changed in a way that files are now in proper packages.
The tests are now in their own path.
Maven dependencies for test has been added.
