--- src/MovieRental.java
@@ -1,17 +1,0 @@
-public class MovieRental {
-    private String movieId;
-    private int days;
-
-    public MovieRental(String movieId, int days) {
-        this.movieId = movieId;
-        this.days = days;
-    }
-
-    public String getMovieId() {
-        return movieId;
-    }
-
-    public int getDays() {
-        return days;
-    }
-}
