+++ src/main/java/com/etraveli/refactoring/model/Movie.java
@@ -1,0 +1,21 @@
+package com.etraveli.refactoring.model;
+
+public class Movie {
+
+  private String title;
+  private String code;
+
+  public Movie(String title, String code) {
+
+    this.title = title;
+    this.code = code;
+  }
+
+  public String getTitle() {
+    return title;
+  }
+
+  public String getCode() {
+    return code;
+  }
+}
