+++ src/main/java/com/etraveli/refactoring/model/MovieRental.java
@@ -1,0 +1,20 @@
+package com.etraveli.refactoring.model;
+
+public class MovieRental {
+
+  private String movieId;
+  private int days;
+
+  public MovieRental(String movieId, int days) {
+    this.movieId = movieId;
+    this.days = days;
+  }
+
+  public String getMovieId() {
+    return movieId;
+  }
+
+  public int getDays() {
+    return days;
+  }
+}
